pyproject.toml
[project]
name = "django-250809"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
]
